+++ source/test-suite/fixtures/FinalResultHaddockExpected.hs
@@ -1,0 +1,8 @@
+module FinalResultHaddockExpected where
+
+convert
+    :: Int
+    -- ^ input value
+    -> Bool
+    -- ^ conversion result
+convert = (> 0)